A few changes
Rambutan is listed as LOW fodmap in a few places if you eat 2 or less so I edited that. I couldn't find enough references to Whey Cheese in relation to the fodmap diet so I excluded it. Generally whey based products are hard to label because the process can be done in many different ways with different ingredients. Also I have quite a few conflicting resources about Wasabi but Wasabi Powder is apparently LOW fodmap so I changed the listed food to be that instead. Thank you for the changes!